Migrate to poise framework
In order to implement code injection, we need to use poise framework which works ontop of serenity. Poise provides a bot command framework that allows us to use modals and other multiline inputs

- Ported existing quiz functionaly to poise library (just using serenity structs inside ApplicationContext)
- Refactored commands to be in a `/commands/` directory vs a single handler file
