+++ Cargo.toml
@@ -8,3 +8,4 @@ edition = "2021"
 serenity = { version="0.11.2", default-features = false, features = ["client", "gateway", "rustls_backend", "model", "collector"] }
 tokio = { version = "1.0", features = ["macros", "rt-multi-thread"] }
 dotenv = { version = "0.15.0" }
+poise = "0.2.1"
