+++ src/question.rs
@@ -8,7 +8,7 @@ use serenity::{
     model::interactions::message_component::ButtonStyle,
 };
 
-#[derive(Debug, PartialEq)]
+#[derive(Debug, PartialEq, Eq)]
 pub enum QuestionTF {
     True,
     False,
